Corrected task 1

+++ 0x1E-search_algorithms/1-binary.c
@@ -37,7 +37,7 @@ int binary_search(int *array, size_t size, int value)
  * @low: index of the lower bound
  * @high: index of the higher bound
  */
-void print_array(int *array, int low, int high)
+void print_array(int *array, size_t low, size_t high)
 {
 	size_t i;
